MOve things around. Remove duplicated functions for stringify
# Conflicts:
#	geoh5py/ui_json/annotations.py
#	geoh5py/ui_json/forms.py
#	geoh5py/ui_json/validations/form.py
