+++ geoh5py/ui_json/ui_json.py
@@ -38,18 +38,22 @@ from pydantic import (
 from geoh5py import Workspace
 from geoh5py.groups import PropertyGroup, UIJsonGroup
 from geoh5py.shared import Entity
-from geoh5py.shared.utils import fetch_active_workspace, str2uuid, stringify
+from geoh5py.shared.utils import (
+    fetch_active_workspace,
+    str2none,
+    str2uuid,
+    stringify,
+)
 from geoh5py.shared.validators import none_to_empty_string
 from geoh5py.ui_json.forms import BaseForm
-from geoh5py.ui_json.validations import ErrorPool, UIJsonError, get_validations
-from geoh5py.ui_json.validations.form import empty_string_to_none
+from geoh5py.ui_json.validation import ErrorPool, UIJsonError, get_validations
 
 
 logger = logging.getLogger(__name__)
 
 OptionalPath = Annotated[
     Path | None,  # pylint: disable=unsupported-binary-operation
-    BeforeValidator(empty_string_to_none),
+    BeforeValidator(str2none),
     PlainSerializer(none_to_empty_string),
 ]
