+++ geoh5py/ui_json/utils.py
@@ -31,7 +31,12 @@ from typing import Any
 from geoh5py import Workspace
 from geoh5py.groups import ContainerGroup, Group
 from geoh5py.objects import ObjectBase
-from geoh5py.shared.utils import fetch_active_workspace
+from geoh5py.shared.utils import (
+    dict_mapper,
+    entity2uuid,
+    fetch_active_workspace,
+    str2none,
+)
 
 
 logger = getLogger(__name__)
@@ -531,3 +536,13 @@ def monitored_directory_copy(
     move(working_path / temp_geoh5, directory_path / temp_geoh5, copy)
 
     return str(directory_path / temp_geoh5)
+
+
+def optional_uuid_mapper(value: Any):
+    """
+    Take values and convert them into UUID or None (or list of).
+
+    :param value: Either a string of entity, or list of.
+    :return: UUID, Nont or list of.
+    """
+    return dict_mapper(value, [str2none, entity2uuid])
