+++ geoh5py/ui_json/validation.py
@@ -20,12 +20,15 @@
 
 from __future__ import annotations
 
+from collections.abc import Callable
 from copy import deepcopy
 from typing import Any, cast
 from uuid import UUID
 from warnings import warn
 
+from geoh5py.data import Data
 from geoh5py.groups import PropertyGroup
+from geoh5py.objects import ObjectBase
 from geoh5py.shared import Entity
 from geoh5py.shared.exceptions import RequiredValidationError
 from geoh5py.shared.validators import (
@@ -315,3 +318,83 @@ class InputValidation:
                 "InputValidators can only be called with dictionary of data or "
                 "(key, value) pair."
             )
+
+
+## Validation utility for UIJson class ##
+class UIJsonError(Exception):
+    """Exception raised for errors in the UIJson object."""
+
+    def __init__(self, message: str):
+        super().__init__(message)
+
+
+class ErrorPool:  # pylint: disable=too-few-public-methods
+    """Stores validation errors for all UIJson members."""
+
+    def __init__(self, errors: dict[str, list[Exception]]):
+        self.pool = errors
+
+    def _format_error_message(self):
+        """Format the error message for the UIJsonError."""
+
+        msg = ""
+        for key, errors in self.pool.items():
+            if errors:
+                msg += f"\t{key}:\n"
+                for i, error in enumerate(errors):
+                    msg += f"\t\t{i}. {error}\n"
+
+        return msg
+
+    def throw(self):
+        """Raise the UIJsonError with detailed list of errors per parameter."""
+
+        message = self._format_error_message()
+        if message:
+            message = "Collected UIJson errors:\n" + message
+            raise UIJsonError(message)
+
+
+def dependency_type_validation(name: str, data: dict[str, Any], params: dict[str, Any]):
+    """Dependency is optional or bool type."""
+
+    dependency = params[name]["dependency"]
+    dependency_form = params[dependency]
+    if "optional" not in dependency_form and not isinstance(data[dependency], bool):
+        raise UIJsonError(
+            f"Dependency {dependency} must be either optional or of boolean type."
+        )
+
+
+def get_validations(form_keys: list[str]) -> list[Callable]:
+    """Returns a list of callable validations based on identifying form keys."""
+    return [VALIDATIONS_MAP[k] for k in form_keys if k in VALIDATIONS_MAP]
+
+
+def mesh_type_validation(name: str, data: dict[str, Any], params: dict[str, Any]):
+    """Promoted value is one of the provided mesh types."""
+
+    mesh_types = params[name]["mesh_type"]
+    obj = data[name]
+    if not isinstance(obj, tuple(mesh_types)):
+        raise UIJsonError(f"Object's mesh type must be one of {mesh_types}.")
+
+
+def parent_validation(name: str, data: dict[str, Any], params: dict[str, Any]):
+    """Data is a child of the parent object."""
+
+    form = params[name]
+    child = data[name]
+    parent = data[form["parent"]]
+
+    if not isinstance(parent, ObjectBase) or (
+        isinstance(child, Data) and parent.get_entity(child.uid)[0] is None
+    ):
+        raise UIJsonError(f"{name} data is not a child of {form['parent']}.")
+
+
+VALIDATIONS_MAP = {
+    "dependency": dependency_type_validation,
+    "mesh_type": mesh_type_validation,
+    "parent": parent_validation,
+}
